requirements.txt
comfy==0.0.1
diffusers==0.27.2
insightface==0.7.3
ipdb==0.13.13
numpy==2.3.3
Pillow==11.3.0
safetensors==0.6.2
torch==2.2.2+cu121
torchvision==0.17.2+cu121
transformers==4.25.1
